display additional message in talent management card if user is nominated in exFeeder program, moved generation of titles and descriptions for list items into view
