+++ services/frontend-v3/src/components/talentManagement/TalentManagement.jsx
@@ -1,34 +1,13 @@
 import React from "react";
 import TalentManagementView from "./TalentManagementView";
-import { FormattedMessage } from "react-intl";
 
 function TalentManagement(props) {
   const data = props.data;
-
-  const getTalentManagementInfo = dataSource => {
-    const locale = localStorage.getItem("lang");
-    const careerMobility = {
-      title: <FormattedMessage id="profile.career.mobility" />,
-      description: dataSource.careerMobility.description[locale] || (
-        <FormattedMessage id="profile.not.specified" />
-      )
-    };
-
-    const talentMatrixResult = {
-      title: <FormattedMessage id="profile.talent.matrix.result" />,
-      description: dataSource.talentMatrixResult.description[locale] || (
-        <FormattedMessage id="profile.not.specified" />
-      )
-    };
-
-    return [careerMobility, talentMatrixResult];
-  };
 
   return (
     <TalentManagementView
       data={data}
       locale={localStorage.getItem("lang")}
-      info={getTalentManagementInfo(data)}
     />
   );
 }
